Revert "Merge branch 'feast-dev:master' into msudhir/add-vector-update-functionality"
This reverts commit 84876789f1f8d0738090bfe7e0cbdbaec8db8f71, reversing
changes made to 0578b9b2bf5efbbc6029e79ecb699b0c51bb931e.
