+++ infra/charts/feast-feature-server/Chart.yaml
@@ -2,7 +2,7 @@ apiVersion: v2
 name: feast-feature-server
 description: Feast Feature Server in Go or Python
 type: application
-version: 0.32.0
+version: 0.31.0
 keywords:
   - machine learning
   - big data
