+++ infra/charts/feast-feature-server/values.yaml
@@ -9,7 +9,7 @@ image:
   repository: feastdev/feature-server
   pullPolicy: IfNotPresent
   # image.tag -- The Docker image tag (can be overwritten if custom feature server deps are needed for on demand transforms)
-  tag: 0.32.0
+  tag: 0.31.0
 
 imagePullSecrets: []
 nameOverride: ""
