+++ infra/charts/feast/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 description: Feature store for machine learning
 name: feast
-version: 0.32.0
+version: 0.31.0
 keywords:
   - machine learning
   - big data
