+++ infra/charts/feast/README.md
@@ -8,7 +8,7 @@ This repo contains Helm charts for Feast Java components that are being installe
    
 ## Chart: Feast
 
-Feature store for machine learning Current chart version is `0.32.0`
+Feature store for machine learning Current chart version is `0.31.0`
 
 ## Installation
 
@@ -65,8 +65,8 @@ See [here](https://github.com/feast-dev/feast/tree/master/examples/java-demo) fo
 | Repository | Name | Version |
 |------------|------|---------|
 | https://charts.helm.sh/stable | redis | 10.5.6 |
-| https://feast-helm-charts.storage.googleapis.com | feature-server(feature-server) | 0.32.0 |
-| https://feast-helm-charts.storage.googleapis.com | transformation-service(transformation-service) | 0.32.0 |
+| https://feast-helm-charts.storage.googleapis.com | feature-server(feature-server) | 0.31.0 |
+| https://feast-helm-charts.storage.googleapis.com | transformation-service(transformation-service) | 0.31.0 |
 
 ## Values
