+++ infra/charts/feast/charts/feature-server/Chart.yaml
@@ -1,8 +1,8 @@
 apiVersion: v1
 description: "Feast Feature Server: Online feature serving service for Feast"
 name: feature-server
-version: 0.32.0
-appVersion: v0.32.0
+version: 0.31.0
+appVersion: v0.31.0
 keywords:
 - machine learning
 - big data
