+++ infra/charts/feast/charts/feature-server/values.yaml
@@ -5,7 +5,7 @@ image:
   # image.repository -- Docker image for Feature Server repository
   repository: feastdev/feature-server-java
   # image.tag -- Image tag
-  tag: 0.32.0
+  tag: 0.31.0
   # image.pullPolicy -- Image pull policy
   pullPolicy: IfNotPresent
