+++ infra/charts/feast/charts/transformation-service/Chart.yaml
@@ -1,8 +1,8 @@
 apiVersion: v1
 description: "Transformation service: to compute on-demand features"
 name: transformation-service
-version: 0.32.0
-appVersion: v0.32.0
+version: 0.31.0
+appVersion: v0.31.0
 keywords:
 - machine learning
 - big data
