+++ java/serving/pom.xml
@@ -244,12 +244,6 @@
     </dependency>
 
     <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-java-sdk-sts</artifactId>
-      <version>1.12.476</version>
-    </dependency>
-
-    <dependency>
       <groupId>com.adobe.testing</groupId>
       <artifactId>s3mock-testcontainers</artifactId>
       <version>2.2.3</version>
