+++ ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@feast-dev/feast-ui",
-  "version": "0.32.0",
+  "version": "0.31.0",
   "private": false,
   "files": [
     "dist"
